auth/requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
greenlet==3.0.3
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==2.1.5
passlib==1.7.4
pyjwt==2.8.0
sqlalchemy==2.0.30
typing-extensions==4.11.0
werkzeug==3.0.3
